Adição do service, repository. Remoção da DAO e Model.

+++ clinicaOdonto/src/main/java/com/dh/clinicaOdonto/repository/IUsuarioRepository.java
@@ -2,6 +2,7 @@ package com.dh.clinicaOdonto.repository;
 
 import com.dh.clinicaOdonto.entity.UsuarioEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-public interface IUsuarioRepository extends JpaRepository<UsuarioEntity, Integer> {
+public interface IUsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
 }
